rename tests and add temp dir
