+++ tests/integration_tests/__init__.py
@@ -4,5 +4,14 @@ from tests.integration_tests.ema_train_integration_test import EMAIntegrationTes
 from tests.integration_tests.lr_test import LRTest
 from tests.integration_tests.pose_estimation_models_test import PoseEstimationModelsIntegrationTest
 from tests.integration_tests.yolo_nas_integration_test import YoloNASIntegrationTest
+from tests.integration_tests.data_adapter.test_dataloader_adapter import DataloaderAdapterTest
+from tests.integration_tests.data_adapter.test_dataloader_adapter_non_regression import DataloaderAdapterNonRegressionTest
 
-__all__ = ["EMAIntegrationTest", "LRTest", "PoseEstimationModelsIntegrationTest", "YoloNASIntegrationTest"]
+__all__ = [
+    "EMAIntegrationTest",
+    "LRTest",
+    "PoseEstimationModelsIntegrationTest",
+    "YoloNASIntegrationTest",
+    "DataloaderAdapterTest",
+    "DataloaderAdapterNonRegressionTest",
+]
